+++ src/test/scala/com/atomist/param/ParameterTest.scala
@@ -45,7 +45,7 @@ class ParameterTest extends FlatSpec with Matchers {
   }
 
   it should "reject a param that doesn't match anchored regexp" in {
-    shouldReject(Parameter("foo","""^[a-z]\w*$"""), Seq("FavoriteColour"))
+    shouldReject(InputParam, Seq("FavoriteColour"))
   }
 
   it should "don't reject parameter that matches a stricter pattern" in {